+++ node/full_node_integration_test.go
@@ -358,7 +358,7 @@ func testSingleAggregatorSingleFullNodeTrustedHash(t *testing.T, source Source) 
 	require.NoError(waitForFirstBlock(node1, source))
 
 	// Get the trusted hash from node1 and pass it to node2 config
-	trustedHash, err := node1.hExService.headerStore.GetByHeight(aggCtx, 1)
+	trustedHash, err := node1.hExService.HeaderStore().GetByHeight(aggCtx, 1)
 	require.NoError(err)
 	node2.conf.TrustedHash = trustedHash.Hash().String()
 	require.NoError(node2.Start())
